More refactoring to inner classes to module level classes.
