+++ android/src/main/java/org/mozilla/osmdroid/tileprovider/modules/MapTileModuleProviderBase.java
@@ -173,19 +173,4 @@ public abstract class MapTileModuleProviderBase implements OSMConstants {
 		}
 	}
 
-    /**
-	 * Thrown by a tile provider module in TileLoader.loadTile() to signal that it can no longer
-	 * function properly. This will typically clear the pending queue.
-	 */
-	public class CantContinueException extends Exception {
-		private static final long serialVersionUID = 146526524087765133L;
-
-		public CantContinueException(final String pDetailMessage) {
-			super(pDetailMessage);
-		}
-
-		public CantContinueException(final Throwable pThrowable) {
-			super(pThrowable);
-		}
-	}
 }
